pyproject.toml
[tool.poetry]
name = "for item in $(cat requirements-txt); do   poetry add \"${item}\"; done"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "for item in $(cat requirements"}]

[tool.poetry.dependencies]
python = "^3.10"
asgiref = "3.5.2"
astroid = "2.11.7"
certifi = "2022.6.15"
charset-normalizer = "2.1.1"
dill = "0.3.5.1"
django = "4.1"
djangorestframework = "3.13.1"
djangorestframework-simplejwt = "5.2.0"
idna = "3.3"
imagekitio = "2.2.8"
isort = "5.10.1"
jinja2 = "3.1.2"
lazy-object-proxy = "1.7.1"
markupsafe = "2.1.1"
mccabe = "0.7.0"
pillow = "9.2.0"
platformdirs = "2.5.2"
pyjwt = "2.4.0"
pylint = "2.14.5"
pytz = "2022.2.1"
redmail = "0.4.1"
requests = "2.28.1"
sqlparse = "0.4.2"
tomli = "2.0.1"
tomlkit = "0.11.4"
urllib3 = "1.26.11"
wrapt = "1.14.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
